requirements-dev.txt
black==19.10b0
eth-brownie>=1.14.4,<2.0.0
